vms-157: widen syscall arg ABI to 64-bit (LLP64 pointer-truncation fix)
Wiring the backend surfaced a latent LLP64 defect: alpha-dec-vms has long=4
but the Linux-Alpha kernel is LP64 (64-bit register args, 64-bit pointers).
musl's generic glue casts every syscall arg through __scc == (long), which on
this port truncated + sign-extended pointer args -- a stack iovec at
0x7743_c8937790 became 0xffffffff_c8937790 and writev() EFAULTed, so user
main's stdout never reached the terminal though main WAS entered.

syscall_arch.h is included before musl's #ifndef __scc fallback, so override
__scc and syscall_arg_t to long long (8B) there, and widen __vms_alpha_syscall
+ its $0/$16-$21 register vars to long long so the trap carries full 64-bit
pointers. This is the port's syscall register width, independent of VMS long.

+++ tools/cross-alpha-vms/musl-arch/src/internal/vms_alpha_syscall.c
@@ -31,15 +31,25 @@
  * cc1-compiled callers (and syscall_cp.s's linkage pair) expect.
  */
 
-long __vms_alpha_syscall(long n, long a1, long a2, long a3, long a4, long a5, long a6)
+/*
+ * Args and result are `long long` (64-bit), NOT `long` -- on this LLP64 port
+ * `long` is only 32 bits, but the Linux-Alpha kernel passes every syscall
+ * argument (and returns pointers) in full 64-bit registers. Truncating a
+ * pointer argument through a 32-bit `long` is exactly the writev-EFAULT bug the
+ * 64-bit __scc / syscall_arg_t in syscall_arch.h exists to prevent; the register
+ * variables here must be equally wide so no truncation reappears in the trap.
+ */
+long long __vms_alpha_syscall(long long n, long long a1, long long a2,
+			      long long a3, long long a4, long long a5,
+			      long long a6)
 {
-	register long r0  __asm__("$0")  = n;   /* syscall number -> result/errno  */
-	register long r16 __asm__("$16") = a1;  /* a0 */
-	register long r17 __asm__("$17") = a2;  /* a1 */
-	register long r18 __asm__("$18") = a3;  /* a2 */
-	register long r19 __asm__("$19") = a4;  /* a3 in; error flag out           */
-	register long r20 __asm__("$20") = a5;  /* a4 */
-	register long r21 __asm__("$21") = a6;  /* a5 */
+	register long long r0  __asm__("$0")  = n;   /* syscall number -> result/errno */
+	register long long r16 __asm__("$16") = a1;  /* a0 */
+	register long long r17 __asm__("$17") = a2;  /* a1 */
+	register long long r18 __asm__("$18") = a3;  /* a2 */
+	register long long r19 __asm__("$19") = a4;  /* a3 in; error flag out          */
+	register long long r20 __asm__("$20") = a5;  /* a4 */
+	register long long r21 __asm__("$21") = a6;  /* a5 */
 
 	__asm__ __volatile__ (
 		"callsys"
